Having Issues with sending time over the socket
I am attempting to I guess create a second socket here to send the timeStamp to the server.  I cant seem to get it working properly.  It is causing an error that says timeSent is an UndefinedProperty.  I do not know if my approach to the problem is correct.  My understanding of sockets might be wrong here.  Anything can help.